refactor(accounting): extract PeakExportService from AccountingService (Wave-4 P3)
Behavior-preserving: exportJournalWithPeakCodes moved verbatim into a new
PeakExportService; AccountingService delegates (signature + controller wiring
unchanged). Read-only, no shared helpers. Existing peak-export tests pass via
the delegating facade (PeakExportService wired into the spec's test module).
