+++ apps/api/src/modules/accounting/accounting.pl-expense.spec.ts
@@ -3,6 +3,7 @@ import { PrismaService } from '../../prisma/prisma.service';
 import { JournalAutoService } from '../journal/journal-auto.service';
 import { CompanyResolverService } from '../journal/company-resolver.service';
 import { AccountingService } from './accounting.service';
+import { PeakExportService } from './peak-export.service';
 
 type GroupRow = { accountCode: string; _sum: { debit: Prisma.Decimal | null; credit: Prisma.Decimal | null } };
 
@@ -16,7 +17,12 @@ function makeService(groupRows: GroupRow[]) {
   const companyResolver = {
     getFinanceCompanyId: jest.fn().mockResolvedValue('finance-co-1'),
   } as unknown as CompanyResolverService;
-  const svc = new AccountingService(prisma, {} as JournalAutoService, companyResolver);
+  const svc = new AccountingService(
+    prisma,
+    {} as JournalAutoService,
+    companyResolver,
+    {} as PeakExportService,
+  );
   return { svc, journalLineGroupBy };
 }
 
@@ -85,7 +91,12 @@ function makeFullService(groupRows: GroupRow[]) {
   const companyResolver = {
     getFinanceCompanyId: jest.fn().mockResolvedValue('finance-co-1'),
   } as unknown as CompanyResolverService;
-  return new AccountingService(prisma, {} as JournalAutoService, companyResolver);
+  return new AccountingService(
+    prisma,
+    {} as JournalAutoService,
+    companyResolver,
+    {} as PeakExportService,
+  );
 }
 
 describe('AccountingService.getProfitLossReport (expense wiring)', () => {
@@ -138,7 +149,12 @@ describe('AccountingService.getMonthlyPLSummary (expense wiring)', () => {
     const companyResolver = {
       getFinanceCompanyId: jest.fn().mockResolvedValue('finance-co-1'),
     } as unknown as CompanyResolverService;
-    return new AccountingService(prisma, {} as JournalAutoService, companyResolver);
+    return new AccountingService(
+      prisma,
+      {} as JournalAutoService,
+      companyResolver,
+      {} as PeakExportService,
+    );
   }
 
   it('includeFinanceExpenses=true: per-month FINANCE expenses subtracted from revenue', async () => {
