+++ apps/api/src/modules/accounting/accounting.service.spec.ts
@@ -1,6 +1,7 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { Prisma } from '@prisma/client';
 import { AccountingService } from './accounting.service';
+import { PeakExportService } from './peak-export.service';
 import { PrismaService } from '../../prisma/prisma.service';
 import { JournalAutoService } from '../journal/journal-auto.service';
 import { CompanyResolverService } from '../journal/company-resolver.service';
@@ -110,6 +111,7 @@ describe('AccountingService', () => {
     const module: TestingModule = await Test.createTestingModule({
       providers: [
         AccountingService,
+        PeakExportService,
         { provide: PrismaService, useValue: prisma },
         { provide: JournalAutoService, useValue: journalAutoService },
         {
